assert: move comparison evaluation to internal
Moving Eval to internal allows other packages, such as poll and a future
x/suite to perform assertions using comparisons.
